Move StackHelper to integration
