+++ mod/src/main/java/org/shsts/tinactory/compat/jei/gui/FluidScreenHandler.java
@@ -9,7 +9,7 @@ import net.minecraftforge.fluids.FluidStack;
 import org.jetbrains.annotations.Nullable;
 import org.shsts.tinactory.core.gui.client.FluidSlot;
 import org.shsts.tinactory.core.gui.client.MenuScreen;
-import org.shsts.tinactory.core.logistics.StackHelper;
+import org.shsts.tinactory.integration.logistics.StackHelper;
 
 @ParametersAreNonnullByDefault
 @MethodsReturnNonnullByDefault
