+++ mod/src/main/java/org/shsts/tinactory/content/gui/ElectricTankMenu.java
@@ -6,7 +6,7 @@ import net.minecraftforge.fluids.FluidStack;
 import org.shsts.tinactory.content.logistics.ElectricTank;
 import org.shsts.tinactory.core.gui.sync.FluidSyncPacket;
 import org.shsts.tinactory.integration.logistics.IFluidTanksHandler;
-import org.shsts.tinactory.core.logistics.StackHelper;
+import org.shsts.tinactory.integration.logistics.StackHelper;
 
 import static org.shsts.tinactory.AllCapabilities.MENU_FLUID_HANDLER;
 import static org.shsts.tinactory.AllMenus.FLUID_SLOT_CLICK;
