+++ mod/src/main/java/org/shsts/tinactory/content/gui/client/LogisticWorkerScreen.java
@@ -30,9 +30,9 @@ import org.shsts.tinactory.core.gui.client.MenuScreen;
 import org.shsts.tinactory.core.gui.client.RenderUtil;
 import org.shsts.tinactory.core.gui.client.StretchImage;
 import org.shsts.tinactory.core.gui.sync.SetMachineConfigPacket;
-import org.shsts.tinactory.core.logistics.StackHelper;
 import org.shsts.tinactory.core.util.ClientUtil;
 import org.shsts.tinactory.core.util.I18n;
+import org.shsts.tinactory.integration.logistics.StackHelper;
 
 import java.util.Collections;
 import java.util.Comparator;
