+++ mod/src/main/java/org/shsts/tinactory/integration/logistics/CombinedFluidTank.java
@@ -15,7 +15,6 @@ import org.shsts.tinactory.api.logistics.IPort;
 import org.shsts.tinactory.api.logistics.IPortFilter;
 import org.shsts.tinactory.api.logistics.IPortNotifier;
 import org.shsts.tinactory.api.logistics.PortType;
-import org.shsts.tinactory.core.logistics.StackHelper;
 import org.slf4j.Logger;
 
 import java.util.Arrays;
